@@ -1,2 +1,3 @@
 """Version number."""
-__version__ = "2024.2.4"
+
+__version__ = "2024.3.4"
